Add minimal app to show bug
Add minimal Container app that uses middleware to post-process 404s

(The example is simplified enough to remove any usefulness the approach
has. The usefulness is for situations where you need to do something
dynamic with a 404 page but where the URL structure for making those
choices is completely different to the URL structure of the app, e.g.
when URLs from a legacy app need special 404 handling)